add mongodb and move index logic to index.js
