+++ package.json
@@ -15,6 +15,7 @@
     "express": "4.13.3",
     "helmet": "^1.3.0",
     "lodash": "^4.13.1",
+    "mongodb": "^2.2.2",
     "nodemon": "^1.9.2",
     "promise": "^7.1.1"
   },
